refactor: rename data_models/chat/ folder to data_models/user_input/.
