+++ client/lib/core/utils/extensions.dart
@@ -6,13 +6,13 @@ import 'package:client/services.dart';
 import 'package:client/styles/app_asset.dart';
 import 'package:client/styles/styles.dart';
 import 'package:data_models/events/event.dart';
-import 'package:data_models/chat/emotion.dart';
+import 'package:data_models/user_input/emotion.dart';
 import 'package:data_models/community/community.dart';
 import 'package:data_models/community/membership.dart';
 import 'package:provider/provider.dart';
 
 export 'package:data_models/events/event.dart';
-export 'package:data_models/chat/emotion.dart';
+export 'package:data_models/user_input/emotion.dart';
 
 extension DateTimeExtension on DateTime {
   /// More references for date formats - https://www.journaldev.com/17899/java-simpledateformat-java-date-format
