+++ client/lib/features/events/features/live_meeting/presentation/widgets/live_meeting_desktop.dart
@@ -31,7 +31,7 @@ import 'package:client/core/widgets/height_constained_text.dart';
 import 'package:client/features/events/presentation/widgets/periodic_builder.dart';
 import 'package:client/core/utils/platform_utils.dart';
 import 'package:data_models/cloud_functions/requests.dart';
-import 'package:data_models/chat/chat.dart';
+import 'package:data_models/user_input/chat.dart';
 import 'package:data_models/events/live_meetings/live_meeting.dart';
 import 'package:data_models/community/membership.dart';
 import 'package:provider/provider.dart';
