+++ data_models/lib/discussion_threads/discussion_thread.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:data_models/cloud_functions/requests.dart';
-import 'package:data_models/chat/emotion.dart';
+import 'package:data_models/user_input/emotion.dart';
 import 'package:data_models/utils/firestore_utils.dart';
 
 part 'discussion_thread.freezed.dart';
