+++ data_models/lib/user_input/chat.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:data_models/chat/emotion.dart';
+import 'package:data_models/user_input/emotion.dart';
 import 'package:data_models/community/membership.dart';
 import 'package:data_models/utils/firestore_utils.dart';
