+++ data_models/lib/user_input/chat_suggestion_data.dart
