+++ data_models/lib/user_input/emotion.dart
