+++ data_models/lib/user_input/emotion.freezed.dart
