+++ firebase/functions/lib/events/live_meetings/get_meeting_chat_suggestion_data.dart
@@ -8,8 +8,8 @@ import '../../utils/infra/firebase_auth_utils.dart';
 import '../../utils/infra/firestore_utils.dart';
 import '../../utils/utils.dart';
 import 'package:data_models/cloud_functions/requests.dart';
-import 'package:data_models/chat/chat.dart';
-import 'package:data_models/chat/chat_suggestion_data.dart';
+import 'package:data_models/user_input/chat.dart';
+import 'package:data_models/user_input/chat_suggestion_data.dart';
 import 'package:data_models/events/event.dart';
 import 'package:data_models/events/live_meetings/meeting_guide.dart';
 import 'package:data_models/community/membership.dart';
